pyproject.toml
[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gptlov"
version = "0.1.0"
description = "GPTLov - Retrieval-augmented chatbot for Lovdata public law datasets"
readme = "README.md"
authors = [{name = "Prosper AI", email = "engineering@prosper.no"}]
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12",
    "tqdm>=4.66",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "joblib>=1.3",
    "openai>=1.53",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "httpx>=0.27",
    "elasticsearch>=8.14"
]

[project.scripts]
gptlov = "gptlov.cli:main"
